Move TaxelApp to app module
